% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fluid.R
\name{connect}
\alias{connect}
\title{Connect to a Fluid server.}
\usage{
connect(ip, port)
}
\arguments{
\item{ip}{A string that contains the internet protocal address of the Fluid
server in dotted decimal format.}

\item{port}{An integer that contains the port number of the Fluid server.}
}
\description{
Connects the current instance of to the Fluid distributed compute server
located at the specified IP address and port number.
}
