% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fluid.R
\name{submitJob}
\alias{submitJob}
\title{Submits a job to the fluid server to run on a distributed cluster}
\usage{
submitJob(func, data)
}
\description{
Submits a job to the fluid server to run on a distributed cluster
}
